requirements.txt
transformers>=4.36.0
torch>=2.2.0
lightgbm>=4.1.0
pandas>=2.1.4
numpy>=1.26.2,<2.0.0
scikit-learn>=1.3.2
pyyaml>=6.0.1
matplotlib>=3.8.2
seaborn>=0.13.0
tqdm>=4.66.1
shap>=0.44.0
streamlit>=1.28.0
plotly>=5.18.0
